+++ client/src/App.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Home from './components/Home';
-import { FlightFinder, Landing, User, FAQ, About, Contact, Navigation, 
-  SignIn, SignUp } from './components/pages';
+import { FlightFinder, Landing, User, FAQ, About, Contact, Navigation} from './components/pages';
+import SignUpPage from './components/pages/SignUp'
+import SignInPage from './components/pages/SignIn';
 import { MenuBar, Footer } from './components';
 
 class App extends Component {
@@ -19,8 +20,8 @@ class App extends Component {
             <Route exact path='/FAQ' component={FAQ} />
             <Route exact path='/About' component={About} />
             <Route exact path='/Contact' component={Contact} />
-            <Route exact path='/signin' component={SignIn} />
-            <Route exact path='/signup' component={SignUp} />
+            <Route exact path='/signin' component={SignInPage} />
+            <Route exact path='/signup' component={SignUpPage} />
           </Switch>
           <Footer />
         </>
