--- client/src/components/LoginLanding/Landing/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Landing = () => (
-  <div>
-    <h1>Landing</h1>
-  </div>
-);
-
-export default Landing;
