+++ client/src/components/pages/PasswordForget/index.js
