+++ client/src/components/pages/Session/withAuthentication.js
