+++ client/src/components/pages/SignIn/index.js
@@ -30,6 +30,7 @@ class SignInFormBase extends Component {
   }
 
   onSubmit = event => {
+    event.preventDefault();
     const { email, password } = this.state;
 
     this.props.firebase
