+++ client/src/components/pages/SignOut/index.js
